test(ai): split prompt tool preparation test suites
Separate tool mapping and tool-choice assertions into dedicated test files and remove the legacy combined test.

Made-with: Cursor
